cordova: add python script for test setup
Running the tests right now involves multiple steps, and not only because of
cordova. These steps are only documented in TESTING.md, but this is
unwelcoming and error prone.

This script tries to automatize that process, in a similar way to
cordova-paramedic, but with some differences:

- This doesn't force you to use a temporary directory and setup
everything each time. This matters because you can save time when
making changes by just re-installing the plugin or the tests plugin.

- This doesn't do the whole thing that cordova paramedic does for
  collecting the test results, although that's not working for me with
  paramedic anyway.

- This, of course, does the things that paramedic wouldn't be able to
  do, that is:

  - Build the native libraries.
  - Copy them to the right place.
  - Run the tests bundling command.

  This is important because it's easy to forget to do a step in an
  edit-compile-run cycle.