add --cargo-build flag to setup.py
This is mostly to build externally in CI because it's easier to setup
cross that way, but can also be used locally by a developer to only
re-build the jni libraries for the particular architecture he's testing
with, instead of rebuilding all of them. Although you need all of them
to be able to build, there is no need for the ones you are not using to
be up-to-date.

This still requires to run the build command for that particular
architecture manually, though, not sure if there should be an option to
choose from the script instead.