update cryptoxide to latest master
I think we can switch to crates.io as soon as we have a release that
compiles for all our targets, in the meantime, this should enable us to
continue doing releases.

Since cryptoxide doesn't expose `secure_memset` anymore, this replaces
it with `zeroize`. Whether this is needed or not, I'm not sure, since
`secure_memset` was just a `for` loop anyway, we could just replace it
with a loop too, but I don't see any problems with `zeroize`.