cleanup plugin api from legacy conversion
These functions are not really used anymore, and while the cost of
maintaining them is not that high, it's still better to have less things
to care about when updating, or when new users aks questions.

This commit focuses mostly on the bindings, mostly because that's where
most of the 'complexity' (duplication) is found, the underlying code is
not really that problematic, but it may be good to remove it too.

In any case, code can be retrieved from the history if necessary.

- walletRestore
- walletRetrieveFunds
- walletConvert
- conversionTransactionsSize
- conversionTransactionsGet
- conversionGetIgnored
- walletPendingTransactions
- pendingTransactionsSize
- pendingTransactionsGet