Cleanup legacy code and update cordova docs (#219)
* mark ffi delete functions as unsafe

* remove wallet rindex module

since it's unused and basically obsolete

* remove legacy unused utxo functionality

* rebuild genesis binary before generating json for cordova

* ignore exist status when deleting the plugins before installing

Helps in the case the script was interrupted after deleting the plugin
but before installing the new version. If there is an error it would
most probably show up in the following install.

* cordova test: rebuild genesis json info before installing

* remove utxo keys from keys.json

* use python3 binary in test script

* remove utxo keys argument from cordova

* remove unused outdated electron/browser cordova platform

* remove outdated getting_started.md file

* update cordova readme

* add updated genesis test vector

* add cbindgen to the cordova requirements

* join the cordova docs in a single readme

* regen wallet.h

* do not require jcli in the ci integration test build

since the tests don't run, there is not much point in regenerating the
assets, and not installing jcli could save some time.